requirements.txt
streamlit
google-generativeai
python-dotenv
requests
fpdf
PyPDF2
python-docx